requirements.txt
# requirements.txt - Backend Dependencies

# FastAPI Framework y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos y ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Manejo de archivos y formularios
python-multipart==0.0.6

# Procesamiento de imágenes
pillow==10.1.0
opencv-python==4.8.1.78

# Google AI Studio
google-generativeai==0.3.2

# Variables de entorno
python-dotenv==1.0.0

# Validación de datos
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilidades adicionales
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Testing (opcional para desarrollo)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging y monitoreo
structlog==23.2.0

# Validación de imágenes (opcional)
python-magic==0.4.27

# Cache opcional
redis==5.0.1
aioredis==2.0.1